LayoutTests/imported/w3c/web-platform-tests/web-animations/responsive/baselineShift-expected.txt


FAIL baselineShift responsive to style changes assert_equals: expected "80px" but got "baseline"
FAIL baselineShift responsive to inherited changes assert_equals: expected "sub" but got "baseline"
